+++ src/main/java/com/github/redouane59/twitter/helpers/AbstractRequestHelper.java
@@ -2,11 +2,15 @@ package com.github.redouane59.twitter.helpers;
 
 import com.github.redouane59.twitter.TwitterClient;
 import com.github.redouane59.twitter.signature.TwitterCredentials;
+import java.io.File;
 import java.io.IOException;
 import java.net.URL;
+import java.util.Map;
 import java.util.concurrent.TimeUnit;
 import lombok.CustomLog;
 import lombok.Getter;
+import okhttp3.Cache;
+import okhttp3.OkHttpClient;
 
 @CustomLog
 @Getter
@@ -42,5 +46,37 @@ public abstract class AbstractRequestHelper {
             LOGGER.severe(e.getMessage());
             Thread.currentThread().interrupt();
         }
+    }
+
+    public OkHttpClient getHttpClient(String url){
+        long   cacheSize = 1024L * 1024 * 1024; // 1go
+        String path      = "../okhttpCache";
+        File   file      = new File(path);
+        return new OkHttpClient.Builder()
+            .addNetworkInterceptor(new CacheInterceptor(this.getCacheTimeoutFromUrl(url)))
+            .cache(new Cache(file, cacheSize))
+            .readTimeout(60, TimeUnit.SECONDS)
+            .connectTimeout(60, TimeUnit.SECONDS)
+            .build();
+    }
+
+    private int getCacheTimeoutFromUrl(String url){
+        int defaultCache = 48;
+        URL cacheUrl = this.getClass().getClassLoader().getResource("cache-config.json");
+        if(cacheUrl==null){
+            LOGGER.severe("cache-config.json file not found in src/main/resources");
+            return defaultCache;
+        }
+        try {
+            Map<String, Integer> map = TwitterClient.OBJECT_MAPPER.readValue(cacheUrl, Map.class);
+            for(Map.Entry<String, Integer> e : map.entrySet()){
+                if(url.contains(e.getKey())){
+                    return e.getValue();
+                }
+            }
+        } catch (IOException e) {
+            LOGGER.severe(e.getMessage());
+        }
+        return defaultCache;
     }
 }
